moving 'Dot' and 'Marking' components to 'Day' directory
