+++ src/calendar/day/dot/style.js
