+++ src/calendar/day/marking/style.js
